Do not fail when `max_workers` is auto-detected
When auto-detecting the value of `max_workers` by not passing it, or passing `None`, it is initialized by the `ProcessPoolExecutor`/`ThreadPoolExecutor` constructor.

In the current state of the source code, that `None` value can be passed to the `BoundedSemaphore` constructor. The problem is that it needs an integer value.

I propose to re-read the auto-detected value from the current instance.